requirements.txt
certifi==2018.4.16
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.4
psycopg2
celery
gunicorn
redis
requests